Before implementing custom renderobject
